Expand stdlib: collection methods, string ops, BTree, File I/O
- Option[T]: add or, or_else methods
- Result[T, E]: add or, or_else methods
- str: add _bytes_index_of, _to_uppercase, _to_lowercase helpers
- String: add index_of, split, replace, to_uppercase, to_lowercase, push_string
- Array[T]: add get, map, filter, fold, any, all, find
- BTreeMap[K, V]: full sorted-array implementation with binary search
- BTreeSet[T]: full sorted-array implementation with binary search
- File I/O: fopen/fclose/fread/fwrite FFI, File struct with open/read_all/write_str/close
- IoError enum with errno mapping, File.open returns Result[File, IoError]
- Reorder prelude: array before string to resolve import dependency

+++ packages/axion-codegen/src/tests.rs
@@ -2545,7 +2545,7 @@ fn main() -> usize
 #[test]
 fn compile_btreemap_new() {
     let src = "\
-fn main() -> i64
+fn main() -> usize
     let m = BTreeMap[i64, i64].new()
     m.len()
 ";
@@ -2556,7 +2556,7 @@ fn main() -> i64
 #[test]
 fn compile_btreeset_new() {
     let src = "\
-fn main() -> i64
+fn main() -> usize
     let s = BTreeSet[i64].new()
     s.len()
 ";
